refactor: Moved import statements for actions to __interface__.py
