+++ src/unipoll_api/routes/v1/workspace.py
@@ -2,7 +2,7 @@
 from typing import Annotated, Literal
 from fastapi import APIRouter, Body, Depends, HTTPException, Query, status
 from unipoll_api import dependencies as Dependencies
-from unipoll_api import actions
+from unipoll_api.actions.__interface__ import WorkspaceActions, MembersActions, PolicyActions, GroupActions, PermissionsActions, PollActions
 from unipoll_api.exceptions.resource import APIException
 from unipoll_api.documents import Workspace, ResourceID, Policy, Member
 from unipoll_api.schemas import WorkspaceSchemas, PolicySchemas, GroupSchemas, MemberSchemas, PollSchemas
@@ -28,7 +28,7 @@ async def get_workspaces():
     The request does not accept any query parameters.
     """
     try:
-        return await actions.WorkspaceActions.get_workspaces()
+        return await WorkspaceActions.get_workspaces()
     except APIException as e:
         raise HTTPException(status_code=e.code, detail=str(e))
 
@@ -49,7 +49,7 @@ async def create_workspace(input_data: WorkspaceSchemas.WorkspaceCreateInput = B
     Returns the created workspace information.
     """
     try:
-        return await actions.WorkspaceActions.create_workspace(input_data=input_data)
+        return await WorkspaceActions.create_workspace(input_data=input_data)
     except APIException as e:
         raise HTTPException(status_code=e.code, detail=str(e))
 
@@ -113,7 +113,7 @@ async def get_workspace(workspace: Workspace = Depends(Dependencies.get_workspac
                     params["include_policies"] = True
                 if "polls" in include:
                     params["include_polls"] = True
-        return await actions.WorkspaceActions.get_workspace(workspace, **params)
+        return await WorkspaceActions.get_workspace(workspace, **params)
 
     except APIException as e:
         raise HTTPException(status_code=e.code, detail=str(e))
@@ -138,7 +138,7 @@ async def update_workspace(workspace: Workspace = Depends(Dependencies.get_works
     Returns the updated workspace.
     """
     try:
-        return await actions.WorkspaceActions.update_workspace(workspace, input_data)
+        return await WorkspaceActions.update_workspace(workspace, input_data)
     except APIException as e:
         raise HTTPException(status_code=e.code, detail=str(e))
 
@@ -158,7 +158,7 @@ async def delete_workspace(workspace: Workspace = Depends(Dependencies.get_works
     Response has no detail.
     """
     try:
-        await actions.WorkspaceActions.delete_workspace(workspace)
+        await WorkspaceActions.delete_workspace(workspace)
         return status.HTTP_204_NO_CONTENT
     except APIException as e:
         raise HTTPException(status_code=e.code, detail=str(e))
@@ -172,7 +172,7 @@ async def delete_workspace(workspace: Workspace = Depends(Dependencies.get_works
             response_model_exclude_unset=True)
 async def get_workspace_members(workspace: Workspace = Depends(Dependencies.get_workspace)):
     try:
-        return await actions.MembersActions.get_members(workspace)
+        return await MembersActions.get_members(workspace)
     except APIException as e:
         raise HTTPException(status_code=e.code, detail=str(e))
 
@@ -185,7 +185,7 @@ async def get_workspace_members(workspace: Workspace = Depends(Dependencies.get_
 async def add_workspace_members(workspace: Workspace = Depends(Dependencies.get_workspace),
                                 member_data: MemberSchemas.AddMembers = Body(...)):
     try:
-        return await actions.MembersActions.add_members(workspace, member_data.accounts)
+        return await MembersActions.add_members(workspace, member_data.accounts)
     except APIException as e:
         raise HTTPException(status_code=e.code, detail=str(e))
 
@@ -198,7 +198,7 @@ async def add_workspace_members(workspace: Workspace = Depends(Dependencies.get_
 async def remove_workspace_member(workspace: Workspace = Depends(Dependencies.get_workspace),
                                   member: Member = Depends(Dependencies.get_member)):
     try:
-        return await actions.MembersActions.remove_member(workspace, member)
+        return await MembersActions.remove_member(workspace, member)
     except APIException as e:
         raise HTTPException(status_code=e.code, detail=str(e))
 
@@ -213,7 +213,7 @@ async def get_workspace_policies(workspace: Workspace = Depends(Dependencies.get
     try:
         account = await Dependencies.get_account(account_id) if account_id else None
         member = await Dependencies.get_member_by_account(account, workspace) if account else None
-        return await actions.PolicyActions.get_policies(resource=workspace, policy_holder=member)
+        return await PolicyActions.get_policies(resource=workspace, policy_holder=member)
     except APIException as e:
         raise HTTPException(status_code=e.code, detail=str(e))
 
@@ -237,7 +237,7 @@ async def set_workspace_policy(workspace: Workspace = Depends(Dependencies.get_w
     Returns the updated workspace.
     """
     try:
-        return await actions.PolicyActions.update_policy(policy, new_permissions=permissions.permissions)
+        return await PolicyActions.update_policy(policy, new_permissions=permissions.permissions)
     except APIException as e:
         raise HTTPException(status_code=e.code, detail=str(e))
 
@@ -249,7 +249,7 @@ async def set_workspace_policy(workspace: Workspace = Depends(Dependencies.get_w
             response_model=PolicySchemas.PermissionList)
 async def get_workspace_permissions():
     try:
-        return await actions.PermissionsActions.get_workspace_permissions()
+        return await PermissionsActions.get_workspace_permissions()
     except APIException as e:
         raise HTTPException(status_code=e.code, detail=str(e))
 
@@ -262,7 +262,7 @@ async def get_workspace_permissions():
             response_model_exclude_none=True)
 async def get_polls(workspace: Workspace = Depends(Dependencies.get_workspace)):
     try:
-        return await actions.PollActions.get_polls(workspace)
+        return await PollActions.get_polls(workspace)
     except APIException as e:
         raise HTTPException(status_code=e.code, detail=str(e))
 
@@ -276,7 +276,7 @@ async def get_polls(workspace: Workspace = Depends(Dependencies.get_workspace)):
 async def create_poll(workspace: Workspace = Depends(Dependencies.get_workspace),
                       input_data: PollSchemas.CreatePollRequest = Body(...)):
     try:
-        return await actions.PollActions.create_poll(workspace, input_data)
+        return await PollActions.create_poll(workspace, input_data)
     except APIException as e:
         raise HTTPException(status_code=e.code, detail=str(e))
 
@@ -288,7 +288,7 @@ async def create_poll(workspace: Workspace = Depends(Dependencies.get_workspace)
             response_model=GroupSchemas.GroupList)
 async def get_groups(workspace: Workspace = Depends(Dependencies.get_workspace)):
     try:
-        return await actions.GroupActions.get_groups(workspace)
+        return await GroupActions.get_groups(workspace)
     except APIException as e:
         raise HTTPException(status_code=e.code, detail=str(e))
 
@@ -302,6 +302,6 @@ async def get_groups(workspace: Workspace = Depends(Dependencies.get_workspace))
 async def create_group(workspace: Workspace = Depends(Dependencies.get_workspace),
                        input_data: GroupSchemas.GroupCreateInput = Body(...)):
     try:
-        return await actions.GroupActions.create_group(workspace, input_data.name, input_data.description)
+        return await GroupActions.create_group(workspace, input_data.name, input_data.description)
     except APIException as e:
         raise HTTPException(status_code=e.code, detail=str(e))
